Move newTestGovernance to staking_manager_test
